consumer/requirements.txt
fastapi==0.95.0
pydantic==1.8.2
uvicorn==0.13.4
gunicorn==20.1.0
pika==1.3.1
appdirs==1.4.4
black==20.8b1
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
flake8==3.8.4
h11==0.11.0
idna==2.10
mccabe==0.6.1
motor==2.3.0
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.6.0
pyflakes==2.2.0
pymongo==3.11.1
regex==2020.11.13
toml==0.10.2
typed-ast==1.4.1
python-dotenv
requests

